# backend/05_APIs_Externas/api_rest/openapi.yaml
openapi: 3.0.3
info:
  title: METGO 3D API REST
  description: API agrometeorológica MVP — Valle de Aconcagua (Chile)
  version: 1.0.0-mvp
servers:
  - url: http://127.0.0.1:8080
    description: Desarrollo local
  - url: https://metgo-api.onrender.com
    description: Producción Render
...
paths:
...
  /api/public/operaciones/{estacion_id}/ventanas:
    get:
      tags: [operaciones]
      summary: Ventanas operacionales de faena (semáforo por actividad)
      description: |
        Serie horaria con semáforo **verde/amarillo/rojo** por actividad
        (tronadura, transporte, izaje) según viento sostenido, ráfaga,
        visibilidad y precipitación. Sitio E8 Mantos Blancos.
      parameters:
        - name: estacion_id
          in: path
          required: true
          schema: { type: string, example: mb_rajo }
        - name: horas
          in: query
          schema: { type: integer, default: 48, maximum: 168 }
      responses:
        "200":
          description: "Serie horaria [{fecha_hora, viento_*, actividades:{tronadura,transporte,izaje}, nivel_global}]"
        "404":
          description: Estación desconocida
        "503":
          description: Servicio Open-Meteo no disponible
  /api/public/operaciones/alertas:
    get:
      tags: [operaciones]
      summary: Alertas por turno de faena (día 07-19 / noche 19-07)
      description: |
        Resumen del próximo turno: actividades bloqueadas (horas en rojo) por
        punto de faena. Para HSE / planificación minera.
      parameters:
        - name: sitio
          in: query
          schema: { type: string, default: mantos_blancos }
        - name: turno
          in: query
          schema: { type: string, enum: [dia, noche], default: dia }
      responses:
        "200":
          description: "{ sitio, turno, desde, hasta, hay_bloqueo, umbrales_aplicados, estaciones[] }"
        "503":
          description: Servicio Open-Meteo no disponible
  /api/public/operaciones/umbrales:
    get:
      tags: [operaciones]
      summary: Umbrales configurables del semáforo operacional
      description: |
        Umbrales efectivos (amarillo, rojo) por actividad. Defaults conservadores
        + override por sitio + `METGO_OP_UMBRALES_JSON` en el entorno de la API.
      parameters:
        - name: sitio
          in: query
          schema: { type: string, default: mantos_blancos }
      responses:
        "200":
          description: "{ sitio, umbrales, unidades, actividades, nota }"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT